Split the mixing into separate files
